requirements.txt
torch
transformers
datasets
pandas
scikit-learn
flask
flask-sqlalchemy
flask-login
werkzeug
newspaper3k
lxml_html_clean
accelerate
huggingface_hub
transformers-interpret
GoogleNews
beautifulsoup4
sentencepiece
protobuf
matplotlib
seaborn